@@ -2,18 +2,18 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <RootNamespace>Faemiyah.BtDamageResolver.Silo</RootNamespace>
     <ServerGarbageCollection>true</ServerGarbageCollection>
     <ConcurrentGarbageCollection>true</ConcurrentGarbageCollection>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
-    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="3.6.5" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.6.5" />
-    <PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="3.6.5" />
-    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="3.6.5" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="7.1.0" />
+    <PackageReference Include="Microsoft.Orleans.Persistence.AdoNet" Version="7.1.0" />
+    <PackageReference Include="Microsoft.Orleans.Sdk" Version="7.1.0" />
+    <PackageReference Include="Microsoft.Orleans.Server" Version="7.1.0" />
   </ItemGroup>
 
   <ItemGroup>
