@@ -1,11 +1,10 @@
 ﻿using Orleans.Services;
 
-namespace Faemiyah.BtDamageResolver.Services.Interfaces
+namespace Faemiyah.BtDamageResolver.Services.Interfaces;
+
+/// <summary>
+/// Client for <see cref="ICommunicationService"/>.
+/// </summary>
+public interface ICommunicationServiceClient : IGrainServiceClient<ICommunicationService>, ICommunicationService
 {
-    /// <summary>
-    /// Client for <see cref="ICommunicationService"/>.
-    /// </summary>
-    public interface ICommunicationServiceClient : IGrainServiceClient<ICommunicationService>, ICommunicationService
-    {
-    }
 }
