@@ -1,11 +1,10 @@
 ﻿using Orleans.Services;
 
-namespace Faemiyah.BtDamageResolver.Services.Interfaces
+namespace Faemiyah.BtDamageResolver.Services.Interfaces;
+
+/// <summary>
+/// Client for <see cref="ILoggingService"/>.
+/// </summary>
+public interface ILoggingServiceClient : IGrainServiceClient<ILoggingService>, ILoggingService
 {
-    /// <summary>
-    /// Client for <see cref="ILoggingService"/>.
-    /// </summary>
-    public interface ILoggingServiceClient : IGrainServiceClient<ILoggingService>, ILoggingService
-    {
-    }
 }
