@@ -9,131 +9,129 @@ using Faemiyah.BtDamageResolver.Services.Interfaces.Enums;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Orleans.Concurrency;
-using Orleans.Core;
 using Orleans.Runtime;
 
-namespace Faemiyah.BtDamageResolver.Services
+namespace Faemiyah.BtDamageResolver.Services;
+
+/// <summary>
+/// Provides stateful logging methods for grains.
+/// </summary>
+[Reentrant]
+public class LoggingService : GrainService, ILoggingService
 {
+    private const int LoggingDelayMilliseconds = 15000; // Check for logs to write 4 times a minute
+
+    private readonly ILogger<LoggingService> _logger;
+    private readonly FaemiyahLoggingOptions _loggingOptions;
+    private readonly LoggingRepository _loggingRepository;
+    private readonly ConcurrentQueue<GameLogEntry> _gameLogEntries;
+    private readonly ConcurrentQueue<PlayerLogEntry> _playerLogEntries;
+    private readonly ConcurrentQueue<UnitLogEntry> _unitLogEntries;
+    private bool _writerActive;
+
     /// <summary>
-    /// Provides stateful logging methods for grains.
+    /// Initializes a new instance of the <see cref="LoggingService"/> class.
     /// </summary>
-    [Reentrant]
-    public class LoggingService : GrainService, ILoggingService
+    /// <param name="logger">The logging interface.</param>
+    /// <param name="clusterOptions">The cluster options.</param>
+    /// <param name="loggingOptions">The logging options.</param>
+    /// <param name="grainId">The grain ID.</param>
+    /// <param name="silo">The silo.</param>
+    /// <param name="loggerFactory">The logger factory.</param>
+    public LoggingService(
+        ILogger<LoggingService> logger,
+        IOptions<FaemiyahClusterOptions> clusterOptions,
+        IOptions<FaemiyahLoggingOptions> loggingOptions,
+        GrainId grainId,
+        Silo silo,
+        ILoggerFactory loggerFactory) : base(grainId, silo, loggerFactory)
     {
-        private const int LoggingDelayMilliseconds = 15000; // Check for logs to write 4 times a minute
+        _logger = logger;
+        _loggingOptions = loggingOptions.Value;
+        _gameLogEntries = new ConcurrentQueue<GameLogEntry>();
+        _playerLogEntries = new ConcurrentQueue<PlayerLogEntry>();
+        _unitLogEntries = new ConcurrentQueue<UnitLogEntry>();
+        _loggingRepository = new LoggingRepository(loggerFactory.CreateLogger<LoggingRepository>(), clusterOptions.Value);
+    }
 
-        private readonly ILogger<LoggingService> _logger;
-        private readonly FaemiyahLoggingOptions _loggingOptions;
-        private readonly LoggingRepository _loggingRepository;
-        private readonly ConcurrentQueue<GameLogEntry> _gameLogEntries;
-        private readonly ConcurrentQueue<PlayerLogEntry> _playerLogEntries;
-        private readonly ConcurrentQueue<UnitLogEntry> _unitLogEntries;
-        private bool _writerActive;
+    /// <inheritdoc />
+    public override Task Start()
+    {
+        _writerActive = true;
+        Task.Run(LogWriteLoop);
+        _logger.LogInformation("{service} running log writing loop.", this.GetType());
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="LoggingService"/> class.
-        /// </summary>
-        /// <param name="logger">The logging interface.</param>
-        /// <param name="clusterOptions">The cluster options.</param>
-        /// <param name="loggingOptions">The logging options.</param>
-        /// <param name="grainId">The grain ID.</param>
-        /// <param name="silo">The silo.</param>
-        /// <param name="loggerFactory">The logger factory.</param>
-        public LoggingService(
-            ILogger<LoggingService> logger,
-            IOptions<FaemiyahClusterOptions> clusterOptions,
-            IOptions<FaemiyahLoggingOptions> loggingOptions,
-            IGrainIdentity grainId,
-            Silo silo,
-            ILoggerFactory loggerFactory) : base(grainId, silo, loggerFactory)
+        return base.Start();
+    }
+
+    /// <inheritdoc />
+    public override Task Stop()
+    {
+        _writerActive = false;
+
+        return base.Stop();
+    }
+
+    /// <inheritdoc />
+    public Task LogGameAction(DateTime timeStamp, string gameId, GameActionType gameActionType, int actionData)
+    {
+        if (_loggingOptions.LogToDatabase)
         {
-            _logger = logger;
-            _loggingOptions = loggingOptions.Value;
-            _gameLogEntries = new ConcurrentQueue<GameLogEntry>();
-            _playerLogEntries = new ConcurrentQueue<PlayerLogEntry>();
-            _unitLogEntries = new ConcurrentQueue<UnitLogEntry>();
-            _loggingRepository = new LoggingRepository(loggerFactory.CreateLogger<LoggingRepository>(), clusterOptions.Value);
+            _gameLogEntries.Enqueue(new GameLogEntry { ActionData = actionData, ActionType = gameActionType, GameId = gameId, TimeStamp = timeStamp });
         }
 
-        /// <inheritdoc />
-        public override Task Start()
-        {
-            _writerActive = true;
-            Task.Run(LogWriteLoop);
-            _logger.LogInformation("{service} running log writing loop.", this.GetType());
+        return Task.CompletedTask;
+    }
 
-            return base.Start();
+    /// <inheritdoc />
+    public Task LogPlayerAction(DateTime timeStamp, string userId, PlayerActionType playerActionType, int actionData)
+    {
+        if (_loggingOptions.LogToDatabase)
+        {
+            _playerLogEntries.Enqueue(new PlayerLogEntry { ActionData = actionData, ActionType = playerActionType, PlayerId = userId, TimeStamp = timeStamp });
         }
 
-        /// <inheritdoc />
-        public override Task Stop()
-        {
-            _writerActive = false;
+        return Task.CompletedTask;
+    }
 
-            return base.Stop();
+    /// <inheritdoc />
+    public Task LogUnitAction(DateTime timeStamp, string unitId, UnitActionType unitActionType, int actionData)
+    {
+        if (_loggingOptions.LogToDatabase)
+        {
+            _unitLogEntries.Enqueue(new UnitLogEntry { ActionData = actionData, ActionType = unitActionType, UnitId = unitId, TimeStamp = timeStamp });
         }
 
-        /// <inheritdoc />
-        public Task LogGameAction(DateTime timeStamp, string gameId, GameActionType gameActionType, int actionData)
+        return Task.CompletedTask;
+    }
+
+    private async Task LogWriteLoop()
+    {
+        while (_writerActive)
         {
-            if (_loggingOptions.LogToDatabase)
+            if (_gameLogEntries.IsEmpty && _playerLogEntries.IsEmpty && _unitLogEntries.IsEmpty)
             {
-                _gameLogEntries.Enqueue(new GameLogEntry { ActionData = actionData, ActionType = gameActionType, GameId = gameId, TimeStamp = timeStamp });
+                _logger.LogDebug("LoggingService has nothing to do, sleeping for {delay} milliseconds.", LoggingDelayMilliseconds);
+                await Task.Delay(LoggingDelayMilliseconds);
             }
-
-            return Task.CompletedTask;
-        }
-
-        /// <inheritdoc />
-        public Task LogPlayerAction(DateTime timeStamp, string userId, PlayerActionType playerActionType, int actionData)
-        {
-            if (_loggingOptions.LogToDatabase)
+            else
             {
-                _playerLogEntries.Enqueue(new PlayerLogEntry { ActionData = actionData, ActionType = playerActionType, PlayerId = userId, TimeStamp = timeStamp });
-            }
-
-            return Task.CompletedTask;
-        }
-
-        /// <inheritdoc />
-        public Task LogUnitAction(DateTime timeStamp, string unitId, UnitActionType unitActionType, int actionData)
-        {
-            if (_loggingOptions.LogToDatabase)
-            {
-                _unitLogEntries.Enqueue(new UnitLogEntry { ActionData = actionData, ActionType = unitActionType, UnitId = unitId, TimeStamp = timeStamp });
-            }
-
-            return Task.CompletedTask;
-        }
-
-        private async Task LogWriteLoop()
-        {
-            while (_writerActive)
-            {
-                if (_gameLogEntries.IsEmpty && _playerLogEntries.IsEmpty && _unitLogEntries.IsEmpty)
+                if (!_gameLogEntries.IsEmpty)
                 {
-                    _logger.LogDebug("LoggingService has nothing to do, sleeping for {delay} milliseconds.", LoggingDelayMilliseconds);
-                    await Task.Delay(LoggingDelayMilliseconds);
+                    _logger.LogDebug("LoggingService writing {count} game log entries.", _gameLogEntries.Count);
+                    await _loggingRepository.WriteGameLogEntries(_gameLogEntries);
                 }
-                else
+
+                if (!_playerLogEntries.IsEmpty)
                 {
-                    if (!_gameLogEntries.IsEmpty)
-                    {
-                        _logger.LogDebug("LoggingService writing {count} game log entries.", _gameLogEntries.Count);
-                        await _loggingRepository.WriteGameLogEntries(_gameLogEntries);
-                    }
+                    _logger.LogDebug("LoggingService writing {count} user log entries.", _playerLogEntries.Count);
+                    await _loggingRepository.WriteUserLogEntries(_playerLogEntries);
+                }
 
-                    if (!_playerLogEntries.IsEmpty)
-                    {
-                        _logger.LogDebug("LoggingService writing {count} user log entries.", _playerLogEntries.Count);
-                        await _loggingRepository.WriteUserLogEntries(_playerLogEntries);
-                    }
-
-                    if (!_unitLogEntries.IsEmpty)
-                    {
-                        _logger.LogDebug("LoggingService writing {count} unit log entries.", _unitLogEntries.Count);
-                        await _loggingRepository.WriteUnitLogEntries(_unitLogEntries);
-                    }
+                if (!_unitLogEntries.IsEmpty)
+                {
+                    _logger.LogDebug("LoggingService writing {count} unit log entries.", _unitLogEntries.Count);
+                    await _loggingRepository.WriteUnitLogEntries(_unitLogEntries);
                 }
             }
         }
